# internal/server/docs/swagger.yaml
definitions:
  entities.Book:
    properties:
      author:
        type: string
      id:
        type: string
      name:
        type: string
      pageCount:
        type: integer
      publisher:
        type: string
      readPage:
        type: integer
      reading:
        type: boolean
      summary:
        type: string
      year:
        type: string
    type: object
  entities.BookChangesResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  entities.BookDetailResponse:
    properties:
      data:
        $ref: '#/definitions/entities.Book'
      status:
        type: string
    type: object
  entities.BooksResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entities.Book'
        type: array
      status:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Bookshelf API
  version: "1.0"
paths:
  /api/v1/books:
    get:
      consumes:
      - application/json
      description: Get all books
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.BooksResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.BookChangesResponse'
      summary: Get all books
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Create book
      parameters:
      - description: Book
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/entities.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.BookChangesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.BookChangesResponse'
      summary: Create book
      tags:
      - books
  /api/v1/books/{bookId}:
    delete:
      consumes:
      - application/json
      description: Delete book
      parameters:
      - description: Book ID
        in: path
        name: bookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.BookChangesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.BookChangesResponse'
      summary: Delete book
      tags:
      - books
    get:
      consumes:
      - application/json
      description: Get book by id
      parameters:
      - description: Book ID
        in: path
        name: bookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.BookDetailResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.BookChangesResponse'
      summary: Get book by id
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Update book
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.BookChangesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.BookChangesResponse'
      summary: Update book
      tags:
      - books
swagger: "2.0"